# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api de location des meubles",
        "description": "il s'agit de la documentation complète de chaque methode,route,etc",
        "contact": {
            "email": "zakiyoubababodi@gmail.com "
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/permission/index": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get all permissions",
                "operationId": "2306a1f461cca99b72b7b4de8fa9ac21",
                "responses": {
                    "200": {
                        "description": "List of permissions"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permission/store": {
            "post": {
                "tags": [
                    "Permission"
                ],
                "summary": "Create a new permission ",
                "operationId": "aaa9a1489978d8237d358f1010932130",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "create"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission  created successfully"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permission/show/{id}": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "Get a specific permission by ID",
                "operationId": "1cae0ef6ddb328b988765a8129134b9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permission",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permission details"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permission/update/{id}": {
            "put": {
                "tags": [
                    "Permission"
                ],
                "summary": "Update a permission by ID",
                "operationId": "10621bd1d3510464e7a1e4e7e3c77dfe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permission",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "permission_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "create"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permission updated successfully"
                    },
                    "404": {
                        "description": "Permission not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/permission/destroy/{id}": {
            "delete": {
                "tags": [
                    "Permission"
                ],
                "summary": "Delete a permission by ID",
                "operationId": "1d26711ac0e6cd7a1fe89efc131e7404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the permission",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Permission deleted successfully"
                    },
                    "404": {
                        "description": "Permission not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}